+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed metrics order in dashSingleValue.vue
+- Fixed error in dashSingleValue.vue on data source changed
 
 ## [1.91.0]